backend/requirements.txt
fastapi 
uvicorn 
python-multipart 
sentence-transformers 
scikit-learn 
python-docx 
pdfplumber
google-generativeai
python-dotenv